+++ denominator-cli/src/test/java/denominator/cli/DenominatorTest.java
@@ -43,6 +43,7 @@ public class DenominatorTest {
                 "mock       mem:mock                                            false          ",
                 "clouddns   https://identity.api.rackspacecloud.com/v2.0        true           password       username password",
                 "clouddns   https://identity.api.rackspacecloud.com/v2.0        true           apiKey         username apiKey",
+                "designate  http://localhost:5000/v2.0                          true           password       tenantId username password",
                 "dynect     https://api2.dynect.net/REST                        false          password       customer username password",
                 "route53    https://route53.amazonaws.com                       true           accessKey      accessKey secretKey",
                 "route53    https://route53.amazonaws.com                       true           session        accessKey secretKey sessionToken",